# project-documentation/api-reference/openapi.yml
openapi: 3.0.3
info:
  title: Drools Rule Engine Microservice API
  description: |
    A high-performance business rule execution microservice built with Spring Boot and Drools.

    ## Features
    - **High Performance**: Support for 100-1000 requests per second
    - **Memory Stable**: Proper resource disposal prevents memory leaks and OOM errors
    - **Real-time Monitoring**: Memory diagnostics with automatic warnings
    - **Security Hardened**: Admin auth, DRL sandboxing, security headers, input validation, rate limiting, CORS
    - **Production Ready**: Health monitoring, metrics, circuit breakers
    - **Scalable**: Multi-tier caching with S3 storage backend

    ## Architecture
    - **Main API**: Rule execution on port 8080
    - **Admin API**: Management and monitoring on port 8080 (same as main API)

    ## Security Features
    - **Admin Authentication**: API key (`X-Admin-API-Key` header) for `/admin/*` endpoints
    - **DRL Sandboxing**: Blocklist-based rule content scanning before compilation
    - **Security Headers**: 7 security headers on all responses (CSP, HSTS, X-Frame-Options, etc.)
    - Input validation with configurable limits
    - Rate limiting with HTTP headers (per remote IP, X-Forwarded-For ignored)
    - Request size protection (including chunked transfer encoding)
    - CORS configuration (empty default; wildcard only in dev profiles)
    - Sensitive data sanitization in logs

    ## Recent Updates (2026-05-11)
...

...

...

...

  version: "1.0.0"
...
servers:
...
paths:
...
  /execute-rule:
    post:
      tags:
        - Rule Execution
      summary: Execute Business Rule
      description: |
        Execute a business rule with provided input data.
        
        **Security Features:**
        - Input validation for rule ID format and data payload
        - Rate limiting per client (configurable)
        - Request size limits (default: 10MB)
        - Timeout protection (default: 30 seconds)
        
        **Performance:**
        - P99 latency < 100ms for cached rules
        - P99 latency < 500ms for cache miss
        - Automatic retry with circuit breaker protection
        
      operationId: executeRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleExecutionRequest'
            examples:
              simple_discount:
                summary: Simple Discount Rule
                value:
                  rule_id: "simple.discount"
                  data:
                    amount: 100
                    customer_type: "regular"
              vip_discount:
                summary: VIP Customer Discount
                value:
                  rule_id: "pricing.discount.vip"
                  data:
                    amount: 150
                    customer_type: "VIP"
                    region: "US"
              complex_pricing:
                summary: Complex Pricing Rule
                value:
                  rule_id: "pricing.dynamic.surge"
                  data:
                    base_amount: 200
                    customer_tier: "gold"
                    time_of_day: "peak"
                    region: "US-WEST"
                    items_count: 5
      responses:
        '200':
          description: Rule executed successfully
          headers:
            X-RateLimit-Limit:
              description: Request limit per minute
              schema:
                type: integer
                example: 1000
            X-RateLimit-Remaining:
              description: Remaining requests in current window
              schema:
                type: integer
                example: 999
            X-RateLimit-Reset:
              description: Unix timestamp when rate limit resets
              schema:
                type: integer
                example: 1753166400
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleExecutionResponse'
              examples:
                success_response:
                  summary: Successful Rule Execution
                  value:
                    rule_id: "simple.discount"
                    result:
                      amount: 90
                      discount: 10
                      discount_percentage: 10
                    error: null
                    execution_time_ms: 15
                    timestamp: "2025-07-22T12:00:00Z"
        '400':
          description: Invalid input or rule execution error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validation_error:
                  summary: Input Validation Error
                  value:
                    rule_id: "invalid.rule"
                    result: null
                    error:
                      code: "INVALID_INPUT"
                      message: "Request validation failed"
                      details: "Rule ID contains invalid characters"
                      timestamp: "2025-07-22T12:00:00Z"
                execution_error:
                  summary: Rule Execution Error
                  value:
                    rule_id: "pricing.discount.vip"
                    result: null
                    error:
                      code: "RULE_EXECUTION_ERROR"
                      message: "Failed to execute rule"
                      details: "Missing required field 'customer_type'"
                      timestamp: "2025-07-22T12:00:00Z"
        '404':
          description: Rule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                rule_not_found:
                  summary: Rule Not Found
                  value:
                    rule_id: "non.existent.rule"
                    result: null
                    error:
                      code: "RULE_NOT_FOUND"
                      message: "Rule not found"
                      details: "Rule 'non.existent.rule' does not exist"
                      timestamp: "2025-07-22T12:00:00Z"
        '408':
          description: Request timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                timeout_error:
                  summary: Request Timeout
                  value:
                    rule_id: "slow.rule"
                    result: null
                    error:
                      code: "TIMEOUT_ERROR"
                      message: "Operation timed out after 30 seconds"
                      details: "Rule execution exceeded configured timeout"
                      timestamp: "2025-07-22T12:00:00Z"
        '413':
          description: Request payload too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                payload_too_large:
                  summary: Request Too Large
                  value:
                    rule_id: null
                    result: null
                    error:
                      code: "REQUEST_TOO_LARGE"
                      message: "Request size exceeds maximum limit"
                      details: "Request size: 11MB, Maximum allowed: 10MB"
                      timestamp: "2025-07-22T12:00:00Z"
        '429':
          description: Rate limit exceeded
          headers:
            X-RateLimit-Limit:
              description: Request limit per minute
              schema:
                type: integer
                example: 1000
            X-RateLimit-Remaining:
              description: Remaining requests (0 when rate limited)
              schema:
                type: integer
                example: 0
            X-RateLimit-Reset:
              description: Unix timestamp when rate limit resets
              schema:
                type: integer
                example: 1753166400
            X-RateLimit-Reset-After:
              description: Seconds until rate limit resets
              schema:
                type: integer
                example: 60
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                rate_limit_exceeded:
                  summary: Rate Limit Exceeded
                  value:
                    rule_id: null
                    result: null
                    error:
                      code: "RATE_LIMIT_EXCEEDED"
                      message: "Rate limit exceeded"
                      details: "Too many requests. Please try again later."
                      timestamp: "2025-07-22T12:00:00Z"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service temporarily unavailable (circuit breaker open)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                service_unavailable:
                  summary: Service Unavailable
                  value:
                    rule_id: null
                    result: null
                    error:
                      code: "SERVICE_UNAVAILABLE"
                      message: "External service 's3' is temporarily unavailable (open)"
                      details: "Please try again later"
                      timestamp: "2025-07-22T12:00:00Z"

  # Admin API Endpoints
...
components:
  schemas:
    RuleExecutionRequest:
      type: object
      required:
        - rule_id
        - data
      properties:
        rule_id:
          type: string
          pattern: '^[a-zA-Z0-9._-]+$'
          maxLength: 255
          description: |
            Unique identifier for the rule to execute.
            
            **Format:**
            - Uses dot notation (e.g., "pricing.discount.vip")
            - Mapped to file path: "pricing/discount/vip.drl"
            - Only alphanumeric, dots, dashes, underscores allowed
            - Maximum 255 characters
          example: "pricing.discount.vip"
        data:
          type: object
          description: |
            Input data for rule execution.
            
            **Validation:**
            - Maximum 100 fields per request (configurable)
            - String values limited to 10,000 characters (configurable)
            - Number values limited to 1,000,000,000 (configurable)
            - No dangerous patterns (script injection protection)
          additionalProperties: true
          example:
            amount: 150
            customer_type: "VIP"
            region: "US"
    RuleExecutionResponse:
      type: object
      properties:
        rule_id:
          type: string
          description: ID of the executed rule
          example: "pricing.discount.vip"
        result:
          type: object
          description: Rule execution result data
          additionalProperties: true
          example:
            amount: 120
            discount: 30
            discount_percentage: 20
        error:
          $ref: '#/components/schemas/ErrorInfo'
        execution_time_ms:
          type: integer
          description: Rule execution time in milliseconds
          example: 15
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
          example: "2025-07-22T12:00:00Z"
    ErrorResponse:
      type: object
      properties:
        rule_id:
          type: string
          nullable: true
          description: Rule ID that caused the error (if applicable)
        result:
          type: object
          nullable: true
          description: Always null for error responses
        error:
          $ref: '#/components/schemas/ErrorInfo'
    ErrorInfo:
      type: object
      properties:
        code:
          type: string
          description: Machine-readable error code
          enum:
            - "RULE_NOT_FOUND"
            - "RULE_EXECUTION_ERROR"
            - "RULE_COMPILATION_ERROR"
            - "INVALID_INPUT"
            - "REQUEST_TOO_LARGE"
            - "RATE_LIMIT_EXCEEDED"
            - "TIMEOUT_ERROR"
            - "SERVICE_UNAVAILABLE"
            - "STORAGE_ERROR"
            - "CACHE_ERROR"
            - "INTERNAL_ERROR"
          example: "RULE_NOT_FOUND"
        message:
          type: string
          description: Human-readable error message
          example: "Rule not found"
        details:
          type: string
          description: Additional error details
          example: "Rule 'pricing.invalid.rule' does not exist"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2025-07-22T12:00:00Z"
...
  securitySchemes:
    AdminApiKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-API-Key
      description: |
        API Key for admin endpoint authentication.

        **Usage:**
        - Required for all `/admin/*` endpoints when `ADMIN_API_KEY` is configured
        - Set via `ADMIN_API_KEY` environment variable
        - When not configured (empty), admin auth is disabled (development mode)
        - Returns 401 Unauthorized if key is missing or invalid

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API Key for client identification and rate limiting.

        **Usage:**
        - Used for client identification in rate limiting
        - Optional but recommended for production
        - Helps with monitoring and analytics

    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer token authentication.

        **Usage:**
        - Alternative client identification method
        - Used in rate limiting and monitoring
        - JWT tokens supported
security:
  - ApiKeyAuth: []
  - BearerAuth: []
  - {}  # Allow unauthenticated access
...